fix: resolve `/quotes` return same records and caching feat (#116)
This PR fixes the issue of /quotes endpoint returning the same records on each request. Also refactored controller for this route to do caching of the total count of records for faster response time. It now returns 5 random records on this endpoint.

Optimizes the auth flow of rate limiting to make sure that all free and protected routes get identified properly.  Updated tests accordingly too.

closes #115